extracted controllers and partials
