+++ app/controllers/peeps.rb
@@ -1,0 +1,7 @@
+class Chitter < Sinatra::Base
+
+  get '/peeps' do
+    erb :'/peeps/peeps'
+  end
+
+end
