+++ app/controllers/users.rb
@@ -1,0 +1,25 @@
+class Chitter < Sinatra::Base
+
+  get '/users/new' do
+    @user = User.new
+    erb :'/users/new'
+  end
+
+  post '/users/create' do
+    @user = User.create(
+      email: params[:email],
+      password: params[:password],
+      password_confirmation: params[:password_confirmation],
+      name: params[:name],
+      username: params[:username]
+    )
+    if @user.save
+      session[:user_id] = @user.id
+      redirect '/peeps'
+    else
+      flash.now[:errors] = @user.errors.full_messages
+      erb :'users/new'
+    end
+  end
+
+end
