+++ app/views/_welcome.erb
@@ -1,0 +1,7 @@
+<% if current_user %>
+  Welcome to chitter <%= current_user.username %>
+  <form method='post' action='/sessions'>
+    <input type='hidden' name='_method' value='delete'>
+    <input type='submit' value='Log out'>
+  </form>
+<% end %>
