+++ app/views/layout.erb
@@ -10,29 +10,8 @@
   </head>
 
   <body>
-    <% if current_user %>
-      Welcome to chitter <%= current_user.username %>
-      <form method='post' action='/sessions'>
-        <input type='hidden' name='_method' value='delete'>
-        <input type='submit' value='Log out'>
-      </form>
-    <% end %>
-
-    <% if flash[:notice] %>
-      <div id='notice'>
-        <%= flash[:notice] %>
-      </div>
-    <% end %>
-
-    <% if flash[:errors] && !flash[:errors].empty? %>
-      Errors:
-      <ul id='errors'>
-        <% flash[:errors].each do |error| %>
-          <li><%= error %></li>
-        <% end %>
-      </ul>
-    <% end %>
-
-    <%= yield %>
+      <%= partial :welcome %>
+      <%= partial :flash %>
+      <%= yield %>
   </body>
 </html>
